requirements.txt
altair==4.2.0
appnope==0.1.3
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.5
bokeh==2.4.3
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
extra-streamlit-components==0.1.56
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter_client==7.4.4
jupyter_core==4.11.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.21.6
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pyarrow==9.0.0
pydeck==0.8.0b4
Pygments==2.13.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.13.0
streamlit-bokeh-events==0.1.2
streamlit-option-menu==0.3.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
wcwidth==0.2.5
zipp==3.10.0
